Moved pubnub to new module
